Rename languageModel.js to matching.js
